data-platform/requirements.txt
dbt-core==1.7.14
dbt-duckdb==1.7.2
streamlit
pandas
